Add exam support for the new design
